Refactor transaction handling by casting gas cost to u64 for balance calculations and updating balance updates in blockchain simulation

+++ crates/kari/src/blockchain_simulation.rs
@@ -108,13 +108,13 @@ pub fn run_blockchain(running: Arc<Mutex<bool>>, miner_address: String) {
             for tx in transactions.iter() {
                 let mut balances = BALANCES.as_mut().unwrap().lock().unwrap();
                 if let Some(sender_balance) = balances.get_mut(&tx.sender) {
-                    *sender_balance -= tx.amount + tx.gas_cost;
+                    *sender_balance -= tx.amount + tx.gas_cost as u64;
                 }
                 *balances.entry(tx.receiver.clone()).or_insert(0) += tx.amount;
             }
 
             // เพิ่มรางวัลให้กับนักขุดจากค่าธรรมเนียมธุรกรรม
-            let transaction_fees: u64 = new_block.transactions.iter().map(|tx| tx.gas_cost).sum();
+            let transaction_fees: u64 = new_block.transactions.iter().map(|tx| tx.gas_cost as u64).sum();
             BALANCES.as_mut().unwrap().lock().unwrap().entry(miner_address.clone()).and_modify(|balance| *balance += transaction_fees + miner_reward).or_insert(transaction_fees + miner_reward);
 
             // Update TOTAL_TOKENS only if it's less than the max supply
